+++ Assets/XChartsDemo/Charts/Bar/Bar20_BarRace.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 using XCharts.Runtime;
 
@@ -15,10 +17,13 @@ namespace XChartsDemo
         {
             chart = gameObject.GetComponent<BarChart>();
             chart.ClearData();
+            nowList.Clear();
             for (int i = 0; i < 5; i++)
             {
+                var value = Random.Range(0, 200);
+                nowList.Add(value);
                 chart.AddYAxisData("y" + i);
-                chart.AddData(0, Random.Range(0, 200));
+                chart.AddData(0, value);
             }
         }
 
@@ -31,16 +36,30 @@ namespace XChartsDemo
             }
         }
 
+        List<int> nowList = new List<int>();
         void UpdateData()
         {
             var serie = chart.GetSerie(0);
-            serie.animation.change.duration = updateTime* 1000;
+            serie.animation.change.duration = updateTime * 1000;
+            RandomIncrement(nowList, Random.Range(1, 200), Random.Range(10, 2000));
             for (int i = 0; i < serie.dataCount; i++)
             {
-                if (Random.Range(0, 1f) > 0.9f)
-                    chart.UpdateData(0, i, chart.GetData(0, i) + Mathf.Round(Random.Range(0, 2000)));
-                else
-                    chart.UpdateData(0, i, chart.GetData(0, i) + Mathf.Round(Random.Range(0, 200)));
+                chart.UpdateData(0, i, nowList[i]);
+            }
+        }
+
+        public static void RandomIncrement(List<int> list, int baseIncrement = 1, int fastIncrement = 2)
+        {
+            int currentMax = list.Max();
+            int fastIndex = Random.Range(0, list.Count);
+
+            for (int i = 0; i < list.Count; i++)
+            {
+                list[i] += (i == fastIndex) ? fastIncrement : baseIncrement;
+            }
+            if (list.All(x => x < currentMax))
+            {
+                list[fastIndex] = currentMax + fastIncrement;
             }
         }
     }