+++ Packages/XCharts/Runtime/Component/Animation/AnimationStyle.cs
@@ -306,7 +306,7 @@ namespace XCharts.Runtime
                 startIndex = anim.context.currPointIndex == paths.Count - 1 ?
                     paths.Count - 2 :
                     anim.context.currPointIndex;
-                if (startIndex < 0 || startIndex > paths.Count - 2) startIndex = 0;
+                if (startIndex < 0 || startIndex >= paths.Count - 1) return;
             }
             else
             {
@@ -336,9 +336,12 @@ namespace XCharts.Runtime
             {
                 return;
             }
-            anim.context.currPoint = sp;
-            anim.context.destPoint = ep;
-            anim.Init(currDetailProgress, totalDetailProgress, paths.Count - 1);
+
+            if (anim.Init(currDetailProgress, totalDetailProgress, paths.Count - 1))
+            {
+                anim.context.currPoint = sp;
+                anim.context.destPoint = ep;
+            }
         }
 
         public bool IsEnd()
@@ -362,12 +365,23 @@ namespace XCharts.Runtime
                 return true;
             var animation = activedAnimation;
             if (animation != null && animation.context.end)
+            {
                 return true;
+            }
             if (IsSerieAnimation())
             {
-                if (m_FadeOut.context.start) return m_FadeOut.context.currProgress <= m_FadeOut.context.destProgress;
-                else if (m_Addition.context.start) return m_Addition.context.currProgress >= m_Addition.context.destProgress;
-                else return m_FadeIn.context.currProgress >= m_FadeIn.context.destProgress;
+                if (m_FadeOut.context.start)
+                {
+                    return m_FadeOut.context.currProgress <= m_FadeOut.context.destProgress;
+                }
+                else if (m_Addition.context.start)
+                {
+                    return m_Addition.context.currProgress >= m_Addition.context.destProgress;
+                }
+                else
+                {
+                    return m_FadeIn.context.currProgress >= m_FadeIn.context.destProgress;
+                }
             }
             else if (IsDataAnimation())
             {
@@ -546,7 +560,7 @@ namespace XCharts.Runtime
         public float GetChangeDuration()
         {
             if (m_Enable && m_Change.enable)
-                return m_Change.duration;
+                return m_Change.context.currDuration > 0 ? m_Change.context.currDuration : m_Change.duration;
             else
                 return 0;
         }
@@ -554,7 +568,7 @@ namespace XCharts.Runtime
         public float GetAdditionDuration()
         {
             if (m_Enable && m_Addition.enable)
-                return m_Addition.duration;
+                return m_Addition.context.currDuration > 0 ? m_Addition.context.currDuration : m_Addition.duration;
             else
                 return 0;
         }
@@ -562,7 +576,7 @@ namespace XCharts.Runtime
         public float GetInteractionDuration()
         {
             if (m_Enable && m_Interaction.enable)
-                return m_Interaction.duration;
+                return m_Interaction.context.currDuration > 0 ? m_Interaction.context.currDuration : m_Interaction.duration;
             else
                 return 0;
         }